+++ src/JsonApi/Hydrator.php
@@ -21,11 +21,18 @@ class Hydrator
     protected $typeMapper;
 
     /**
-     * @param \Swis\JsonApi\Client\Interfaces\TypeMapperInterface $typeMapper
+     * @var \Swis\JsonApi\Client\JsonApi\LinksParser
      */
-    public function __construct(TypeMapperInterface $typeMapper)
+    protected $linksParser;
+
+    /**
+     * @param \Swis\JsonApi\Client\Interfaces\TypeMapperInterface $typeMapper
+     * @param \Swis\JsonApi\Client\JsonApi\LinksParser            $linksParser
+     */
+    public function __construct(TypeMapperInterface $typeMapper, LinksParser $linksParser)
     {
         $this->typeMapper = $typeMapper;
+        $this->linksParser = $linksParser;
     }
 
     /**
@@ -41,6 +48,10 @@ class Hydrator
 
         if ($jsonApiItem->has('attributes')) {
             $item->fill($jsonApiItem->get('attributes')->asArray(true));
+        }
+
+        if ($jsonApiItem->has('links')) {
+            $item->setLinks($this->linksParser->parse($jsonApiItem->get('links')->asArray(false)));
         }
 
         if ($jsonApiItem->has('meta')) {